% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{d_err}
\alias{d_err}
\title{D-error}
\usage{
d_err(par, design, n_alts)
}
\arguments{
\item{par}{Vector containing parameter values.}

\item{design}{Design matrix.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}
}
\value{
D-error.
}
\description{
Function to calculate d error given a design, and parameter values.
}

