% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{hessian}
\alias{hessian}
\title{Hessian}
\usage{
hessian(par, design, covar, n_alts)
}
\arguments{
\item{par}{parameter value(s)}

\item{design}{The design matrix}

\item{covar}{the covariance matrix}

\item{n_alts}{The number of alternatives in each choice set}
}
\value{
the hessian matrix
}
\description{
Hessian
}

