% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{imp_sampling}
\alias{imp_sampling}
\title{Importance sampling}
\usage{
imp_sampling(prior_mode, prior_covar, design, n_alts, Y, ...)
}
\arguments{
\item{prior_mode}{Vector containing the mode of the multivariate normal distribution which is the prior.}

\item{prior_covar}{The covariance matrix of the multivariate normal prior distribution.}

\item{design}{A design matrix in which each row is a profile.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{Y}{A binary response vector.}
}
\value{
A list containing samples and their associated weights, which together represent the posterior distribution.
}
\description{
This functions samples from an imortance density (multivariate t-distribution),
and gives weightes to the samples according to the posterior distribution. The prior is
a normal distribution.
}

