% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{info_design}
\alias{info_design}
\title{Fisher Information of design}
\usage{
info_design(par, design, n_alts)
}
\arguments{
\item{par}{A vector containing the parameter values}

\item{design}{A design matrix. Each row is a profile, successive rows form choice sets.}
}
\value{
Fisher Information matrix.
}
\description{
Returns the Fisher Information of a design, given parameter values
}

