% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{info_set}
\alias{info_set}
\title{Fisher Information of choice set}
\usage{
info_set(par, c_set)
}
\arguments{
\item{par}{A vector containing the parameter values}

\item{c_set}{A matrix representing the choice set. Each row is a profile.}
}
\value{
Fisher Information matrix
}
\description{
Return the fisher information matrix of a particular choice set,
given parameter values.
}

