% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_mod_fed.R
\name{KL}
\alias{KL}
\title{KL information}
\usage{
KL(set, par_samples, weights)
}
\arguments{
\item{par_samples}{A matrix in which each row is a sample.}

\item{weights}{A vector containing the weights of all the samples.}

\item{Matrix}{in which each row is an alternative of the choice set.}
}
\value{
The Kullback-Leibler divergence
}
\description{
Calculates the Kullback-Leibler divergence for a choice set, given parameter values.
}

