% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_mod_fed.R, R/efficiency.R
\name{KL_info}
\alias{KL_info}
\title{KL set selecting}
\usage{
KL_info(lvls, par_samples, weights, n_alts)

KL_info(lvls, par_samples, weights, n_alts)
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{par_samples}{A matrix in which each row is a sample.}

\item{weights}{A vector containing the weights of all the samples.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{n_sets}{Numeric value indicating the number of choice sets.}

\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{par_samples}{A matrix in which each row is a sample from a multivariate prior parameter distribution.}

\item{weights}{A vector containing the weights of all the samples.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}
}
\value{
The most efficient choice set

The most efficient choice set, based on the KL information criterion.
}
\description{
Provides the set that maximizes the Kullback-Leibler divergence, given parameter values.

Calcultates the most effecient choice set by maximizing the Kullback-Leibler distance.
}

