% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_mod_fed.R
\name{KL_select}
\alias{KL_select}
\title{KL set selecting}
\usage{
KL_select(lvls, n_sets, n_alts, par_samples, weights)
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{n_sets}{Numeric value indicating the number of choice sets.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{par_samples}{A matrix in which each row is a sample.}

\item{weights}{A vector containing the weights of all the samples.}
}
\value{
The most efficient choice set
}
\description{
Provides the set that maximizes the Kullback-Leibler divergence, given parameter values.
}

