% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{lattice}
\alias{lattice}
\title{Lattice standard normal generation.}
\usage{
lattice(K, b = 2, m = 9)
}
\arguments{
\item{K}{Numeric value indicating the dimensionality of the grid.}

\item{b}{Numeric value indicating the base.}

\item{m}{Numeric value. N=b^m.}
}
\value{
Matrix of lattice points drawn from a multivariate standard normal distribution.
}
\description{
Generates a grid of points (N=b^m), coming from a multivariate standard normal distribution with dim = K.
}

