% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{lattice_mvt}
\alias{lattice_mvt}
\title{Lattice t-distribution generation.}
\usage{
lattice_mvt(mode, cvar, df, ...)
}
\arguments{
\item{K}{Numeric value indicating the dimensionality of the distribution.}

\item{b}{Numeric value indicating the base.}

\item{m}{Numeric value. N=b^m.}
}
\value{
Matrix of lattice points drawn from a multivariate t-distribution.
}
\description{
Generates a grid of points (N=b^m), coming from a multivariate t-distribution.
}

