% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{Lik}
\alias{Lik}
\title{Likelihood function}
\usage{
Lik(par, design, n_alts, Y)
}
\arguments{
\item{par}{parameter value(s)}

\item{design}{The design matrix}

\item{n_alts}{The number of alternatives in each choice set}

\item{Y}{the response vector}
}
\value{
the likelihood
}
\description{
Likelihood function
}

