% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{logPost}
\alias{logPost}
\title{log Posterior}
\usage{
logPost(parameters, prior_mode, prior_covar, design, n_alts, Y)
}
\arguments{
\item{prior_mode}{vector containing the prior mode}

\item{prior_covar}{matrix containing the prior covariance}

\item{design}{The Design matrix}

\item{n_alts}{The number of alternatives in each choice set}

\item{Y}{the response vector}

\item{par}{parameter value(s) for wich the logposterior is calculated}
}
\value{
the logposterior probability
}
\description{
Calculates the logposterior with a normal prior density
}

