% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_sampling.R
\name{modes}
\alias{modes}
\title{modes}
\usage{
modes(samples, weights, s)
}
\arguments{
\item{samples}{Matrix with samples coming from the posterior distribution. Each row is a sample.}

\item{weights}{Vector containing the weights of the samples.}
}
\value{
The mode of the posterior distribution.
}
\description{
This functions returns a vector with the posterior modes,
given samples from the posterior and their weights (importance sampling).
}

