% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{present_Gil}
\alias{present_Gil}
\title{label design matrix}
\usage{
present_Gil(design, levels, lvl_names, n_alts)
}
\arguments{
\item{design}{An effect coded design matrix.}

\item{levels}{vector containing the number of levels for each attribute.}

\item{lvl_names}{A list containing the values of each level of each attribute.}

\item{n_alts}{Number of alternatives per choice set.}
}
\value{
A desing matrix with presentable attributelevels.
}
\description{
Transforms an effect coded design matrix into the design matrix containing the real attributelevels.
This design can be used to present to respondents.
}

