% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{profiles}
\alias{profiles}
\title{Profile generation}
\usage{
profiles(lvls, intercept = FALSE, contr = "contr.sum")
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{intercept}{Logical argument indicating whether an intercept should be included. The Default is False.}

\item{contr}{See model.matrix for more information. The default is effect coded.}
}
\value{
A matrix containig all possible profiles.
}
\description{
Function to generate all possible combinations of attribute levels.
}

