% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{savedrop}
\alias{savedrop}
\title{Save to dropbox}
\usage{
savedrop(des, Y, dir, filename)
}
\arguments{
\item{des}{A design matrix in which each row is a profile.}

\item{Y}{A response vector.}
}
\description{
Save design and responses to dropbox map
}

