% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DB_mod_fed.R
\name{DB_seq_fed}
\alias{DB_seq_fed}
\title{DB sequential set}
\usage{
DB_seq_fed(des, lvls, n_alts, par_samples, weights, prior_covar, ...)
}
\arguments{
\item{des}{A design matrix in which each row is a profile.}

\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{par_samples}{A matrix in which each row is a sample from the multivariate prior parameter distribution.}

\item{weights}{A vector containing the weights of the samples.}

\item{prior_covar}{covariance matrix of the prior distribution.}
}
\value{
The most DB efficient choice set
}
\description{
Adds the most DB efficient choice set to a design, given (updated) parameter values.
}

