% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{design}
\alias{design}
\title{Design generation}
\usage{
design(lvls, n_sets, n_alts, intercept = FALSE, contr = "contr.sum")
}
\arguments{
\item{n_sets}{Numeric value indicating the number of choide sets.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{profiles}{Matrix containing (all) possible profiles.}
}
\value{
A design matrix
}
\description{
Function to generate a Design matrix given the number of choice sets,
the number of alternatives per choice set and a matrix containting possible profiles.
}

