% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{full_sets}
\alias{full_sets}
\title{All choice sets}
\usage{
full_sets(cand, n_alts, intercept = F, contr = "contr.sum")
}
\arguments{
\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{lvls}{A vector which contains for each attribute, the number of levels.}
}
\value{
Matrix with all possible combinations of profiles.
}
\description{
Generates all possible combinations of choice sets, given all candidate profiles
and the number of alternatives.
}

