% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{full_sets}
\alias{full_sets}
\title{All choice sets}
\usage{
full_sets(lvls, n_alts, mindiff, intercept = F, contr = "contr.sum")
}
\arguments{
\item{lvls}{A vector which contains for each attribute, the number of levels.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}

\item{mindiff}{The minimal number of atrribute levels that needs to be different in a choice set.}

\item{intercept}{logical value indicating whether an intercept should be present.}
}
\value{
Matrix with all possible combinations of profiles, taking into account the mindiff constraint.
}
\description{
Generates all possible combinations of choice sets,
taking into account the minimum ammount of attributelevels that need to be different per choice set.
}

