% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{KL_info}
\alias{KL_info}
\title{Kullback-Leibler information criterion}
\usage{
KL_info(fp, fcomb, par_samples, weights, n_alts)
}
\arguments{
\item{fp}{A Matrix containing every possible profile (combination of attributelevels).}

\item{fcomb}{A matrix containing in each row the to be considered profile combinations (choice sets).
Generated by full_sets.}

\item{par_samples}{A matrix in which each row is a sample from a multivariate parameter distribution.}

\item{weights}{A vector containing the weights of all the samples.}

\item{n_alts}{Numeric value indicating the number of alternatives per choice set.}
}
\value{
The most efficient choice set, based on the KL information criterion.
}
\description{
Calcultates the most effecient choice set by maximizing the Kullback-Leibler distance.
}

