% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.R
\name{respond}
\alias{respond}
\title{Response generation}
\usage{
respond(par, design, n_alts)
}
\arguments{
\item{par}{Vector containing parameter values.}

\item{design}{Design matrix.}
}
\value{
Vector with binary responses.
}
\description{
Function to generate responses given parameter values and a design matrix. Chooses the alternative with max probability.
}

