% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{savedrop}
\alias{savedrop}
\title{Save to dropbox}
\usage{
savedrop(d, Y, dir, filename)
}
\arguments{
\item{d}{Matrix containing the design.}

\item{Y}{A response vector.}
}
\description{
Save design and responses to dropbox map
}

